Sync to upstream sources 2014-10-12

+++ src/lisp/ppc-arch.c
@@ -117,7 +117,7 @@ arch_remove_breakpoint(void *pc, unsigned long orig_inst)
     os_flush_icache((os_vm_address_t) pc, sizeof(unsigned long));
 }
 
-static unsigned long *skipped_break_addr, displaced_after_inst;
+static unsigned int *skipped_break_addr, displaced_after_inst;
 static sigset_t orig_sigmask;
 
 void
@@ -133,7 +133,7 @@ arch_do_displaced_inst(os_context_t * scp, unsigned long orig_inst)
     *pc = orig_inst;
     os_flush_icache((os_vm_address_t) pc, sizeof(unsigned int));
 
-    skipped_break_addr = (unsigned long) pc;
+    skipped_break_addr = pc;
 
     /*
      * Replace the next instruction with a
